+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Remove outdated calls to `weldx.asdf.util.get_highest_tag_version` in
   `TimeSeries` and `SpatialData` converters \[{pull}`831`\].
 - Use Ruff in pre-commit-action \[{pull}`824`\].
+- Use MyST-NB for documentation building\[ {pull}`830`\].
 
 ## 0.6.2 (07.11.2022)